+++ .names.sh
@@ -5,6 +5,7 @@ cmp_argh
 cmp_args
 cmp_clap
 cmp_clap3
+cmp_clap4
 cmp_commander
 cmp_flood_tide
 cmp_getopts