+++ bin/bench-release-no-run.sh
@@ -8,6 +8,7 @@ for nm in $NAMES; do
   echo "$nm :::"
   ( cd "comps/$nm" && {
     env -u CARGO_PROFILE_RELEASE_OPT_LEVEL -u CARGO_PROFILE_RELEASE_CODEGEN_UNITS -u CARGO_PROFILE_RELEASE_PANIC \
+      CARGO_PROFILE_BENCH_LTO="fat" \
       cargo bench --bench bench-$nm-curl --bench bench-$nm-one --no-run
       #
       #-u RUSTC_WRAPPER