+++ bin/size-release.one.sh
@@ -7,5 +7,5 @@ for nm in $NAMES; do
   #echo "$A :::"
   a="$a \"target/release/$nm-one\""
 done
-eval size "$a" | rust-gsub -e "^(.*)target/release/([^ ]+)\$" -f "\$1\$2" \
+eval size "$a" | aki-gsub -e "^(.*)target/release/([^ ]+)\$" -f "\$1\$2" \
   > z.size-release.one.log